remote: do Basic or Bearer auth depending on server offer
I don't like this way; its too clever and also awkward because really we
want to be doing this discovery _before_ creating the HttpWrapper, but
its not structured right for that.

Even if it was right, it misses a couple of bits. It doesn't handle the
server offering multiple auth mechanisms and .. something else I forgot.
No matter, I'm gonna redo it.